backend-ui-detection/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
ultralytics>=8.0.0
Pillow>=10.0.0
numpy>=1.24.0
huggingface_hub>=0.20.0
